requirements.txt
gym==0.19.0
h5py==2.10.0
Keras==2.0.6
matplotlib==3.4.1
numpy==1.19.5
pandas==1.2.3
tensorflow==1.14.0
pytorch==1.4.0
tensorboardx==2.4